backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.2
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.10.1
APScheduler==3.10.4
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
httpx==0.28.1
yfinance==1.2.0
openai==1.93.0
tavily-python>=0.7.0
python-dotenv
# Optional enrichment sources (add API keys to .env to enable)
# FRED_API_KEY      → https://fred.stlouisfed.org/docs/api/api_key.html (free)
# FINNHUB_API_KEY   → https://finnhub.io/ (free tier available)
# ALPHA_VANTAGE_API_KEY → https://www.alphavantage.co/ (free tier available)
alembic>=1.13.0
